requirements.txt
streamlit~=1.32.2
pandas>=1.3.0
scikit-learn~=1.4.1.post1
scikit-image~=0.22.0
altair>=4.0
horde-sdk~=0.8.3
Pillow~=10.2.0
python-dotenv~=1.0.1
numpy~=1.26.4
matplotlib==3.8.3
scipy~=1.12.0
requests~=2.31.0
pyperclip~=1.8.2